refactor: move favorite to v2
