--- src/app/favorite/mod.rs
@@ -1,4 +1,0 @@
-pub mod api;
-pub mod model;
-pub mod response;
-pub mod service;
