+++ src/app/mod.rs
@@ -1,6 +1,5 @@
 pub mod article;
 pub mod comment;
-pub mod favorite;
 pub mod follow;
 pub mod healthcheck;
 pub mod tag;
