+++ src/appv2/features/favorite/entities.rs
