+++ src/appv2/features/favorite/mod.rs
@@ -1,0 +1,4 @@
+pub mod controllers;
+pub mod entities;
+pub mod presenters;
+pub mod services;
