+++ src/appv2/features/mod.rs
@@ -1,2 +1,3 @@
+pub mod favorite;
 pub mod profile;
 pub mod user;
